Added updating

+++ utils/deleteFile.js
@@ -1,6 +1,7 @@
 const fs = require("fs");
 
 const deleteFile = (filePath) => {
+  if (!filePath) return;
   fs.unlink(filePath, (err) => {
     if (err) throw err;
   });
